+++ src/components/IconButtons/ButtonBookmark.jsx
@@ -1,8 +1,8 @@
 import { useEffect, useState } from "react";
 import axios from "axios";
-import styles from "../button/Button.module.css";
+import styles from "./Button.module.css"
 import { useRecoilValue } from "recoil";
-import { loginState } from "../../state/LoginState";
+import { loginState } from "../../util/state/LoginState";
 
 const BOOKMARK_COLOR = "#FFC436";
