+++ src/components/IconButtons/ButtonRecommend.jsx
@@ -2,9 +2,9 @@
 // import { loginState } from "../state/LoginState";
 import { useEffect, useState } from "react";
 import axios from "axios";
-import styles from "../button/Button.module.css";
+import styles from "./Button.module.css";
 import { useRecoilValue } from "recoil";
-import { loginState } from "../../state/LoginState";
+import { loginState } from "../../util/state/LoginState";
 
 const RECOMMEND_COLOR = "#337CCF";
